refactoring: duplicates and testscript class removed
