+++ TestShell/TestShell.vcxproj
@@ -139,7 +139,6 @@
       <AdditionalOptions Condition="'$(Configuration)|$(Platform)'=='Debug|x64'"> %(AdditionalOptions)</AdditionalOptions>
       <AdditionalOptions Condition="'$(Configuration)|$(Platform)'=='Release|x64'"> %(AdditionalOptions)</AdditionalOptions>
     </ClCompile>
-    <ClCompile Include="testscript.cpp" />
     <ClCompile Include="testscript_runner.cpp" />
     <ClCompile Include="shell.cpp" />
     <ClCompile Include="test_shell_test.cpp" />
